nydus-image: fix wrongly calculated prefetch table size
Nid in prefetch table is of u32, so prefetch table should be
files_count * size_of::<u32>()

+++ rafs/src/metadata/layout/v6.rs
@@ -1673,7 +1673,7 @@ impl RafsV6PrefetchTable {
 
     /// Get content size of the inode prefetch table.
     pub fn size(&self) -> usize {
-        self.len() * size_of::<u64>()
+        self.len() * size_of::<u32>()
     }
 
     /// Get number of entries in the prefetch table.
